new queries all tests pass
